Inline TFLib into TFModel
We have no intention to pursue this line of work
so merge the two files to compile tests slightly
faster.
